backend/pyproject.toml
[project]
name = "orchestrai-backend"
version = "0.1.0"
description = "OrchestrAI backend: agent run tracking, replay, and evaluation"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "alembic>=1.13",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "redis>=5.0",
  "celery>=5.3",
  "httpx>=0.27",
  "opentelemetry-api>=1.23",
  "opentelemetry-sdk>=1.23",
  "opentelemetry-exporter-otlp>=1.23",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
